Extract schema types from schema.ts
